continuing the work

+++ src/services/categories.services.ts
@@ -4,11 +4,19 @@ import { createCategoryBodyInterfaces } from "../interfaces/categories.interface
 
 @injectable()
 export class CategoriesServices{
-    async createCategory(body: createCategoryBodyInterfaces){
+    async createCategory(body: createCategoryBodyInterfaces, userId: number){
+        body.userId = userId
         return await prisma.category.create({data: body})
     }
 
-    async deleteCategory(categoryId: string){
+    async deleteCategory(categoryId: string, userId: number){
+        if(await prisma.category.findFirst({
+            where: {
+                id: Number(categoryId),
+                AND: {userId: userId}
+            }
+        })){
             return await prisma.category.delete({where: {id: Number(categoryId)}})
+        }
     }
 }
